+++ specs/unit/patterns/behavioral/mediator.spec.js
@@ -8,7 +8,6 @@ describe('mediator', function() {
   var constructor;
   var Mediator;
   var mediator;
-
   beforeEach(function() {
     participantA = jasmine.createSpy('A');
     participantB = jasmine.createSpy('B');
@@ -34,6 +33,16 @@ describe('mediator', function() {
     mediator.broadcast('anonymous broadcast');
     mediator.send('lost', 'participantA', 'non-existant');
     mediator.register('non-existant', participantC);
+  });
+  it('should allow empty options', function() {
+    var emptyOptions = undefined;
+    var Mediator = mediatorBuilder(emptyOptions).build();
+    var mediator = new Mediator();
+    mediator.register('from', (arg) => arg);
+    mediator.register('to', (message) => {
+      expect(message).toEqual('message');
+    });
+    mediator.send('message', 'from', 'to');
   });
   it('should wrap with pattern', function() {
     expect(constructor).toHaveBeenCalledWith('test');