+++ specs/unit/patterns/creational/factory.spec.js
@@ -6,7 +6,6 @@ describe('Factory', function() {
   var CarFactory;
   var Honda, Civic, civic;
   var Toyota, Corolla, corolla;
-
   beforeEach(function() {
     Civic = class {
       constructor(year) {
@@ -39,6 +38,14 @@ describe('Factory', function() {
     civic = Honda.create('civic', 2016);
     corolla = Toyota.create('corolla', 2015);
   });
+  it('should allow empty options', function() {
+    var emptyOptions = undefined;
+    var Factory = factoryBuilder(emptyOptions).build();
+    var factory = new Factory();
+    factory.add('product', (arg) => ({test: arg}));
+    var product = factory.create('product', 'testing');
+    expect(product.test).toEqual('testing');
+  });
   it('should create a Factory', function() {
     expect(CarFactory).toBeDefined();
     expect(Honda).toBeDefined();
@@ -59,7 +66,6 @@ describe('Factory', function() {
   it('should be able to define static properties/functionality', function() {
     expect(CarFactory.test).toBeDefined();
   });
-
   describe('Mid Level: Singleton Factory', function() {
     var SingletonFactory;
     var Constructor;
@@ -80,7 +86,6 @@ describe('Factory', function() {
       expect(singletonFactory.param).toEqual('test');
     });
   });
-
   describe('Advanced Level: Factory of factories', function() {
     var FactoryOfFactories;
     var factoryOfFactories;