+++ src/patterns/behavioral/chainOfResponsibility.js
@@ -19,9 +19,10 @@ export default createPatternBuilder(options => {
     run(...args) {
       for(var i = 0; i < this.chain.length; i++) {
         let next = false;
-        let val = this.chain[i].apply(null, [() => {
-          next = true;
-        }].concat(args));
+        let val = this.chain[i].apply(null, [
+          () => { next = true; },
+          ...args
+        ]);
         if(!next) {
           return val;
         }